SRB2: cakes
git-svn-id: svn://code.srb2.org/SRB2/trunk@5054 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/v_video.c
@@ -235,7 +235,7 @@ static void CV_Gammaxxx_ONChange(void)
 // --------------------------------------------------------------------------
 // Copy a rectangular area from one bitmap to another (8bpp)
 // --------------------------------------------------------------------------
-void VID_BlitLinearScreen(const UINT8 *srcptr, UINT8 *destptr, INT32 width, INT32 height, INT32 srcrowbytes,
+void VID_BlitLinearScreen(const UINT8 *srcptr, UINT8 *destptr, INT32 width, INT32 height, size_t srcrowbytes,
 	size_t destrowbytes)
 {
 	if (srcrowbytes == destrowbytes)
